Fixes for new gyro filtering changes

+++ firmware/src/imu.c
@@ -22,9 +22,9 @@ void imu_init(void)
 void imu_update_from_gyro(struct gyro_data *gyro)
 {
   // Convert the gyro readings to radians and multiply by the time since the last update.
-  float gyro_x = (float)gyro->x * 0.00122173f / 3300.f;
-  float gyro_y = (float)gyro->y * 0.00122173f / 3300.f;
-  float gyro_z = (float)gyro->z * 0.00122173f / 3300.f;
+  float gyro_x = gyro->x * 0.00122173f / 3300.f;
+  float gyro_y = gyro->y * 0.00122173f / 3300.f;
+  float gyro_z = gyro->z * 0.00122173f / 3300.f;
 
   // Create quaternions for each axis rotation.
   Quaternion q_x, q_y, q_z;
